antigravity/pyproject.toml
[project]
name = "antigravity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.31.1",
    "litellm>=1.83.0",
]
